Added dynamic moving header for the subreddit screen for better one-handed UX
A little bit of animation has been added to the top header when scrolling down. The subreddit title will sliding in from the bottom once the top header has been scrolled off-screen.

The top header will probably need additional work, like opacity (with a gradient perhaps), if any images are to be included in the bg